Testing out door features. Added a utility library that allows us to manipulate the box mesh we use for doors. Added contants for door width. Added a door entity to enable easy creation of door-blocks. Moved the camera bit closer.
Changed how contactfilters coming from category filters should not disable contact if both bodies do not have a category filters.

Refactored MapSystem to read the comments on the legacy maps easier, and changed switch to if-else to utilize the short-ranges. Added a static lookup for the VIE tile types.

Changed how shapes are registered on the view side a bit, to avoid a NullPointerException (and added the two newest FLAG and DOOR)